refactor: split history contracts and MCP tests
